requirements.txt
requests
beautifulsoup4
openai
langchain
langchain-openai
langgraph
python-dotenv
streamlit
